feat(test): add backend_test_fixture for improved backup testing
